FEniCS backend: Move variables code
