+++ tlm_adjoint/fenics/variables.py
@@ -1,17 +1,21 @@
-"""This module includes functionality for interacting with FEniCS variables
-and Dirichlet boundary conditions.
+"""FEniCS variables.
 """
 
-from .backend import backend_Constant, backend_ScalarType
+from .backend import (
+    backend_Constant, backend_Function, backend_ScalarType, backend_Vector)
 from ..interface import (
-    SpaceInterface, VariableInterface, space_comm, var_comm, var_dtype,
-    var_is_cached, var_is_static, var_linf_norm, var_lock_state,
-    var_scalar_value, var_space_type)
+    SpaceInterface, VariableInterface, check_space_types,
+    register_subtract_adjoint_derivative_action,
+    subtract_adjoint_derivative_action_base, space_comm, space_id, var_axpy,
+    var_comm, var_copy, var_dtype, var_is_cached, var_is_static, var_linf_norm,
+    var_lock_state, var_new, var_scalar_value, var_space_type)
 
 from ..caches import Caches
+from ..equations import Conversion
 
-from .expr import Replacement, Zero
+from .expr import Replacement, Zero, r0_space
 
+import functools
 import numbers
 import numpy as np
 import weakref
@@ -19,13 +23,17 @@ import weakref
 __all__ = \
     [
         "Constant",
+        "Function",
 
         "ZeroConstant",
+        "ZeroFunction",
 
         "ReplacementConstant",
         "ReplacementFunction",
         "ReplacementZeroConstant",
-        "ReplacementZeroFunction"
+        "ReplacementZeroFunction",
+
+        "to_fenics"
     ]
 
 
@@ -286,6 +294,211 @@ class Constant(backend_Constant):
         self._tlm_adjoint__var_interface_attrs.d_setitem("cache", cache)
 
 
+class FunctionSpaceInterface(SpaceInterface):
+    def _comm(self):
+        return self._tlm_adjoint__space_interface_attrs["comm"]
+
+    def _dtype(self):
+        return backend_ScalarType
+
+    def _id(self):
+        return self._tlm_adjoint__space_interface_attrs["id"]
+
+    def _new(self, *, name=None, space_type="primal", static=False,
+             cache=None):
+        return Function(self, name=name, space_type=space_type, static=static,
+                        cache=cache)
+
+
+def check_vector(fn):
+    @functools.wraps(fn)
+    def wrapped_fn(self, *args, **kwargs):
+        def space_sizes(self):
+            dofmap = self.function_space().dofmap()
+            n0, n1 = dofmap.ownership_range()
+            return (n1 - n0, dofmap.global_dimension())
+
+        def vector_sizes(self):
+            return (self.vector().local_size(), self.vector().size())
+
+        space = self.function_space()
+        vector = self.vector()
+        if vector_sizes(self) != space_sizes(self):
+            raise RuntimeError("Unexpected vector size")
+
+        return_value = fn(self, *args, **kwargs)
+
+        if self.function_space() is not space:
+            raise RuntimeError("Unexpected space")
+        if self.vector() is not vector:
+            raise RuntimeError("Unexpected vector")
+        if vector_sizes(self) != space_sizes(self):
+            raise RuntimeError("Unexpected vector size")
+
+        return return_value
+    return wrapped_fn
+
+
+class FunctionInterface(VariableInterface):
+    def _space(self):
+        return self._tlm_adjoint__var_interface_attrs["space"]
+
+    def _space_type(self):
+        return self._tlm_adjoint__var_interface_attrs["space_type"]
+
+    def _id(self):
+        return self._tlm_adjoint__var_interface_attrs["id"]
+
+    def _name(self):
+        return self.name()
+
+    def _state(self):
+        return self._tlm_adjoint__var_interface_attrs["state"][0]
+
+    def _update_state(self):
+        self._tlm_adjoint__var_interface_attrs["state"][0] += 1
+
+    def _is_static(self):
+        return self._tlm_adjoint__var_interface_attrs["static"]
+
+    def _is_cached(self):
+        return self._tlm_adjoint__var_interface_attrs["cache"]
+
+    def _caches(self):
+        if "caches" not in self._tlm_adjoint__var_interface_attrs:
+            self._tlm_adjoint__var_interface_attrs["caches"] = Caches(self)
+        return self._tlm_adjoint__var_interface_attrs["caches"]
+
+    @check_vector
+    def _zero(self):
+        self.vector().zero()
+
+    @check_vector
+    def _assign(self, y):
+        if isinstance(y, numbers.Real):
+            if len(self.ufl_shape) != 0:
+                raise ValueError("Invalid shape")
+            self.assign(backend_Constant(y))
+        elif isinstance(y, backend_Function):
+            if space_id(y.function_space()) != space_id(self.function_space()):
+                raise ValueError("Invalid function space")
+            self.vector().zero()
+            self.vector().axpy(1.0, y.vector())
+        else:
+            raise TypeError(f"Unexpected type: {type(y)}")
+
+    @check_vector
+    def _axpy(self, alpha, x, /):
+        if isinstance(x, backend_Function):
+            if space_id(x.function_space()) != space_id(self.function_space()):
+                raise ValueError("Invalid function space")
+            self.vector().axpy(alpha, x.vector())
+        else:
+            raise TypeError(f"Unexpected type: {type(x)}")
+
+    @check_vector
+    def _inner(self, y):
+        if isinstance(y, backend_Function):
+            if space_id(y.function_space()) != space_id(self.function_space()):
+                raise ValueError("Invalid function space")
+            return y.vector().inner(self.vector())
+        else:
+            raise TypeError(f"Unexpected type: {type(y)}")
+
+    @check_vector
+    def _linf_norm(self):
+        return self.vector().norm("linf")
+
+    @check_vector
+    def _local_size(self):
+        return self.vector().local_size()
+
+    @check_vector
+    def _global_size(self):
+        return self.function_space().dofmap().global_dimension()
+
+    @check_vector
+    def _local_indices(self):
+        return slice(*self.function_space().dofmap().ownership_range())
+
+    @check_vector
+    def _get_values(self):
+        return self.vector().get_local().copy()  # copy likely not required
+
+    @check_vector
+    def _set_values(self, values):
+        self.vector().set_local(values)
+        self.vector().apply("insert")
+
+    @check_vector
+    def _new(self, *, name=None, static=False, cache=None,
+             rel_space_type="primal"):
+        y = var_copy(self, name=name, static=static, cache=cache)
+        y.vector().zero()
+        space_type = var_space_type(self, rel_space_type=rel_space_type)
+        y._tlm_adjoint__var_interface_attrs.d_setitem("space_type", space_type)  # noqa: E501
+        return y
+
+    @check_vector
+    def _copy(self, *, name=None, static=False, cache=None):
+        y = self.copy(deepcopy=True)
+        if name is not None:
+            y.rename(name, "a Function")
+        if cache is None:
+            cache = static
+        y._tlm_adjoint__var_interface_attrs.d_setitem("space_type", var_space_type(self))  # noqa: E501
+        y._tlm_adjoint__var_interface_attrs.d_setitem("static", static)
+        y._tlm_adjoint__var_interface_attrs.d_setitem("cache", cache)
+        return y
+
+    def _replacement(self):
+        if "replacement" not in self._tlm_adjoint__var_interface_attrs:
+            count = self._tlm_adjoint__var_interface_attrs["replacement_count"]
+            if isinstance(self, Zero):
+                self._tlm_adjoint__var_interface_attrs["replacement"] = \
+                    ReplacementZeroFunction(self, count=count)
+            else:
+                self._tlm_adjoint__var_interface_attrs["replacement"] = \
+                    ReplacementFunction(self, count=count)
+        return self._tlm_adjoint__var_interface_attrs["replacement"]
+
+    def _is_replacement(self):
+        return False
+
+    def _is_scalar(self):
+        return False
+
+    def _is_alias(self):
+        return "alias" in self._tlm_adjoint__var_interface_attrs
+
+
+class Function(backend_Function):
+    """Extends the DOLFIN `Function` class.
+
+    :arg space_type: The space type for the :class:`.Function`. `'primal'`,
+        `'dual'`, `'conjugate'`, or `'conjugate_dual'`.
+    :arg static: Defines whether the :class:`.Function` is static, meaning that
+        it is stored by reference in checkpointing/replay, and an associated
+        tangent-linear variable is zero.
+    :arg cache: Defines whether results involving the :class:`.Function` may be
+        cached. Default `static`.
+
+    Remaining arguments are passed to the DOLFIN `Function` constructor.
+    """
+
+    def __init__(self, *args, space_type="primal", static=False, cache=None,
+                 **kwargs):
+        if space_type not in {"primal", "conjugate", "dual", "conjugate_dual"}:
+            raise ValueError("Invalid space type")
+        if cache is None:
+            cache = static
+
+        super().__init__(*args, **kwargs)
+        self._tlm_adjoint__var_interface_attrs.d_setitem("space_type", space_type)  # noqa: E501
+        self._tlm_adjoint__var_interface_attrs.d_setitem("static", static)
+        self._tlm_adjoint__var_interface_attrs.d_setitem("cache", cache)
+
+
 class ZeroConstant(Constant, Zero):
     """A :class:`.Constant` which is flagged as having a value of zero.
 
@@ -301,6 +514,22 @@ class ZeroConstant(Constant, Zero):
         var_lock_state(self)
         if var_linf_norm(self) != 0.0:
             raise RuntimeError("ZeroConstant is not zero-valued")
+
+
+class ZeroFunction(Function, Zero):
+    """A :class:`.Function` which is flagged as having a value of zero.
+
+    Arguments are passed to the :class:`.Function` constructor, together with
+    `static=True` and `cache=True`.
+    """
+
+    def __init__(self, *args, **kwargs):
+        Function.__init__(
+            self, *args, **kwargs,
+            static=True, cache=True)
+        var_lock_state(self)
+        if var_linf_norm(self) != 0.0:
+            raise RuntimeError("ZeroFunction is not zero-valued")
 
 
 class ReplacementConstant(Replacement):
@@ -331,6 +560,20 @@ class ReplacementZeroFunction(ReplacementFunction, Zero):
         Zero.__init__(self)
 
 
+def to_fenics(y, space, *, name=None):
+    """Convert a variable to a DOLFIN `Function`.
+
+    :arg y: A variable.
+    :arg space: The space for the return value.
+    :arg name: A :class:`str` name.
+    :returns: The DOLFIN `Function`.
+    """
+
+    x = Function(space, space_type=var_space_type(y), name=name)
+    Conversion(x, y).solve()
+    return x
+
+
 def define_var_alias(x, parent, *, key):
     if x is not parent:
         if "alias" in x._tlm_adjoint__var_interface_attrs:
@@ -350,3 +593,44 @@ def define_var_alias(x, parent, *, key):
                 "cache", var_is_cached(parent))
             x._tlm_adjoint__var_interface_attrs.d_setitem(
                 "state", parent._tlm_adjoint__var_interface_attrs["state"])
+
+
+def subtract_adjoint_derivative_action_backend_constant_vector(x, alpha, y):
+    if hasattr(y, "_tlm_adjoint__function"):
+        check_space_types(x, y._tlm_adjoint__function)
+
+    if len(x.ufl_shape) == 0:
+        value = y.max()
+    else:
+        value = np.zeros_like(x.values()).flatten()
+        y_fn = backend_Function(r0_space(x))
+        y_fn.vector().axpy(1.0, y)
+        for i, y_fn_c in enumerate(y_fn.split(deepcopy=True)):
+            value[i] = y_fn_c.vector().max()
+        value.shape = x.ufl_shape
+        value = backend_Constant(value)
+
+    y = var_new(x)
+    y.assign(value)
+    var_axpy(x, -alpha, y)
+
+
+def subtract_adjoint_derivative_action_backend_function_vector(x, alpha, y):
+    if hasattr(y, "_tlm_adjoint__function"):
+        check_space_types(x, y._tlm_adjoint__function)
+
+    if x.vector().local_size() != y.local_size():
+        raise ValueError("Invalid function space")
+    x.vector().axpy(-alpha, y)
+
+
+register_subtract_adjoint_derivative_action(
+    (backend_Constant, backend_Function), object,
+    subtract_adjoint_derivative_action_base,
+    replace=True)
+register_subtract_adjoint_derivative_action(
+    backend_Constant, backend_Vector,
+    subtract_adjoint_derivative_action_backend_constant_vector)
+register_subtract_adjoint_derivative_action(
+    backend_Function, backend_Vector,
+    subtract_adjoint_derivative_action_backend_function_vector)
